PHRAS-2651_event-to-message_4.1 !!! WIP !!!
